Extracted TokenSelector into a new component
