+++ src/components/TradeWidget/TokenRow.tsx
@@ -1,11 +1,10 @@
-import React, { CSSProperties, useMemo } from 'react'
+import React from 'react'
 import { Link } from 'react-router-dom'
 import styled from 'styled-components'
-import Select from 'react-select'
-import { FormatOptionLabelContext } from 'react-select/src/Select'
 import { useFormContext } from 'react-hook-form'
 
 import TokenImg from 'components/TokenImg'
+import TokenSelector from 'components/TokenSelector'
 import { TokenDetails, TokenBalanceDetails } from 'types'
 import { formatAmount, formatAmountFull } from 'utils'
 
@@ -21,23 +20,6 @@ const TokenImgWrapper = styled(TokenImg)`
   height: 4em;
 
   margin-right: 1em;
-`
-
-const SelectBox = styled.div`
-  display: flex;
-  flex-direction: column;
-  align-items: stretch;
-
-  margin: 0em 1em;
-
-  label {
-    text-transform: uppercase;
-    padding-left: 8px; // to align with Select input padding
-  }
-
-  input {
-    margin-left: 0; // to fix extra space on Select search box
-  }
 `
 
 const InputBox = styled.div`
@@ -89,51 +71,6 @@ const WalletDetail = styled.div`
   }
 `
 
-function renderOptionLabel(token: TokenDetails): React.ReactNode {
-  return (
-    <div
-      style={{
-        display: 'flex',
-        alignItems: 'center',
-      }}
-    >
-      <TokenImgWrapper
-        src={token.image}
-        alt={token.name}
-        style={{
-          margin: '0.25em 2em 0.25em 1em',
-        }}
-      />
-      <div
-        style={{
-          display: 'flex',
-          flexDirection: 'column',
-        }}
-      >
-        <div>
-          <strong>{token.symbol}</strong>
-        </div>
-        <div>{token.name}</div>
-      </div>
-    </div>
-  )
-}
-
-function formatOptionLabel(
-  options: { token: TokenDetails },
-  labelMeta: { context: FormatOptionLabelContext },
-): React.ReactNode {
-  const { token } = options
-  const { context } = labelMeta
-  return context === 'value' ? <strong>{token.symbol}</strong> : renderOptionLabel(token)
-}
-
-const customSelectStyles = {
-  control: (provided: CSSProperties): CSSProperties => ({ ...provided, border: 'none' }),
-  menu: (provided: CSSProperties): CSSProperties => ({ ...provided, minWidth: '300px' }),
-  valueContainer: (provided: CSSProperties): CSSProperties => ({ ...provided, minWidth: '4.5em' }),
-}
-
 function displayBalance(balance: TokenBalanceDetails | undefined | null, key: string): string {
   if (!balance) {
     return '0'
@@ -171,8 +108,6 @@ const TokenRow: React.FC<Props> = ({
   isDisabled,
   validateMaxAmount,
 }) => {
-  const options = useMemo(() => tokens.map(token => ({ token, value: token.symbol, label: token.name })), [tokens])
-
   const { register, errors, setValue, watch } = useFormContext()
   const error = errors[inputId]
 
@@ -199,23 +134,13 @@ const TokenRow: React.FC<Props> = ({
   return (
     <Wrapper>
       <TokenImgWrapper alt={selectedToken.name} src={selectedToken.image} />
-      <SelectBox>
-        <label>{selectLabel}</label>
-        <Select
-          isSearchable
-          isDisabled={isDisabled}
-          styles={customSelectStyles}
-          noOptionsMessage={(): string => 'No results'}
-          formatOptionLabel={formatOptionLabel}
-          options={options}
-          value={{ token: selectedToken }}
-          onChange={(selected, { action }): void => {
-            if (action === 'select-option' && 'token' in selected) {
-              onSelectChange(selected.token)
-            }
-          }}
-        />
-      </SelectBox>
+      <TokenSelector
+        label={selectLabel}
+        isDisabled={isDisabled}
+        tokens={tokens}
+        selected={selectedToken}
+        onChange={onSelectChange}
+      />
       <InputBox>
         <input
           className={className}
